+++ docs/markdown/changelog.md
@@ -10,6 +10,7 @@
 - Added support model for text insertion. Special thanks to Anurag Bansal.
 - Created an `open_page()` function to avoid internal duplication. This basically is a wrapper around `FPDF_LoadPage()`, for page index validation.
 - Improved the cedits section in the Readme.
+- Made automatic buffer closure optional for `open_pdf_buffer()`.
 
 
 ## 1.7.0 (2022-05-02)